+++ src/render/backends/gl/qdemonrenderbackendgl3.h
@@ -59,27 +59,27 @@ public:
     ~QDemonRenderBackendGL3Impl() override;
 
 public:
-    quint32 getDepthBits() const override;
-    quint32 getStencilBits() const override;
+    qint32 getDepthBits() const override;
+    qint32 getStencilBits() const override;
     void generateMipMaps(QDemonRenderBackendTextureObject to,
                          QDemonRenderTextureTargetType::Enum target,
                          QDemonRenderHint::Enum genType) override;
 
     void setMultisampledTextureData2D(QDemonRenderBackendTextureObject to,
                                       QDemonRenderTextureTargetType::Enum target,
-                                      size_t samples,
+                                      qint32 samples,
                                       QDemonRenderTextureFormats::Enum internalFormat,
-                                      size_t width,
-                                      size_t height,
+                                      qint32 width,
+                                      qint32 height,
                                       bool fixedsamplelocations) override;
 
     void setTextureData3D(QDemonRenderBackendTextureObject to,
                           QDemonRenderTextureTargetType::Enum target,
-                          quint32 level,
+                          qint32 level,
                           QDemonRenderTextureFormats::Enum internalFormat,
-                          size_t width,
-                          size_t height,
-                          size_t depth,
+                          qint32 width,
+                          qint32 height,
+                          qint32 depth,
                           qint32 border,
                           QDemonRenderTextureFormats::Enum format,
                           const void *hostPtr = nullptr) override;
